first draft new task concept
