+++ src/pseudo_label/pseudolabel_detectron.py
@@ -26,8 +26,8 @@ from math import ceil
 import torch
 
 from utils_asl import load_yaml
-from task import TaskCreator
-from task import TaskCreator
+from task import TaskGenerator
+from task import TaskGenerator
 from datasets_asl import get_dataset
 
 ASL = os.path.join( str(Path.home()), "ASL" )
@@ -46,7 +46,7 @@ else:
 
 # SETUP DATALOADERS
 
-tc = TaskCreator(**exp['task_generator'] )
+tc = TaskGenerator(**exp['task_generator'] )
 
 use_tc = True
 if use_tc:
